More cleaning
Added a script to automatically copy the image variable location from the PRU image map file to an include file for the main processor program. Streamlined the main loop to get rid of the conversion to HSV. Added an RGB implementation but it doesn't give the best image quality, so sticking with UYUV for now. Put in flag to block PRU from writing out next frame until main loop has consumed the current one.